+++ ReleaseNotes.md
@@ -38,6 +38,7 @@
 * **\[BREAKING]** The `Day::getHashCode` method renamed to `Day::hashCode`.
 * **\[BREAKING]** The `Session::getHashCode` method renamed to `Session::hashCode`.
 * **\[BREAKING]** The `DXEndpoint::getEventTypes()` method now can throw exceptions.
+* **\[BREAKING]** Most of the input string arguments in class methods have been replaced with a `StringLike` wrapper for unification.
 * Fixed the naming of endpoints.
 
 ## v4.3.1